Intent: Consolidate CLI token parsing to core helpers Operators involved: [Resonance|Coherence] Affected invariants: [#4] Key changes: [link CLI to tnfr.token_parser, remove redundant exports, drop equivalence test] Expected risks/dissonances: Shared parser import paths adjusted; watch for missing re-exports in CLI Metrics: No metric impact; CLI parsing delegates to core TOKEN_MAP Equivalence map: tnfr.cli._parse_tokens -> tnfr.token_parser._parse_tokens
