+++ src/tnfr/cli/__init__.py
@@ -16,8 +16,7 @@ from .arguments import (
     _add_metrics_parser,
     _args_to_dict,
 )
-from .token_parser import _parse_tokens
-from ..token_map import TOKEN_MAP
+from ..token_parser import _parse_tokens, validate_token, parse_thol, TOKEN_MAP
 from .execution import (
     build_basic_graph,
     apply_cli_config,
@@ -47,7 +46,9 @@ __all__ = (
     "resolve_program",
     "_build_graph_from_args",
     "_load_sequence",
+    "validate_token",
     "_parse_tokens",
+    "parse_thol",
     "TOKEN_MAP",
     "_save_json",
     "_args_to_dict",
