--- src/tnfr/cli/token_parser.py
@@ -1,6 +1,0 @@
-from __future__ import annotations
-
-from ..token_parser import validate_token, _parse_tokens
-from ..token_map import TOKEN_MAP, parse_thol
-
-__all__ = ("validate_token", "_parse_tokens", "parse_thol", "TOKEN_MAP")
